moved shallowUntilTarget into fabrics common folder
