+++ packages/utilities/src/index.ts
@@ -36,7 +36,6 @@ export * from './properties';
 export * from './resources';
 export * from './rtl';
 export * from './scroll';
-export * from './shallowUntilTarget';
 export * from './string';
 export * from './styled';
 export * from './warn';
